% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theory.R
\name{fortenum}
\alias{fortenum}
\title{Forte number from set class}
\usage{
fortenum(set)
}
\arguments{
\item{set}{Numeric vector of pitch-classes in the set}
}
\value{
Character string in the form "n-x" where n is the number of notes
in the set and x is the ordinal position in Forte's list.
}
\description{
Given a pitch-class set (in 12edo only), look up Forte 1973's catalog
number for the corresponding set class.
}
\examples{
fortenum(c(0, 4, 7))
fortenum(c(0, 3, 7))
fortenum(c(4, 8, 11))
}
