% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howfree.R
\name{howfree}
\alias{howfree}
\title{Count a scale's degrees of freedom}
\usage{
howfree(set, ineqmat = NULL, edo = 12, rounder = 10)
}
\arguments{
\item{set}{Numeric vector of pitch-classes in the set}

\item{ineqmat}{Defaults to \code{NULL}, in which case the function assumes
you want to use the standard hyperplane arrangement of MCT. But you can
enter a different matrix that contains normal vectors for any hyperplane
arrangement in the same format as the standard ineqmats.}

\item{edo}{Number of unit steps in an octave. Defaults to \code{12}.}

\item{rounder}{Numeric (expected integer) representing tolerance:
number of decimal places to round to.}
}
\value{
Single nonnegative integer
}
\description{
Some scalar structures can vary their specific pitches much more flexibly
than others while retaining the same overall "color." For instance, the
meantone family of diatonic scales is generated by a line of fifths and
can only vary along one dimension: the size of the generating fifth.
This literally defines a line in the MCT geometry, and if the scale moves
off that line it ceases to have the same structure as the diatonic scale.
(Notably, it stops being non-degenerate well-formed.) By contrast, the
5-limit just diatonic scale is defined by two distinct parameters: the
size of its major third and the size of its perfect fifth. See "Modal
Color Theory," pp. 26-27, for more discussion.
}
\examples{
c_natural_minor <- c(0,2,3,5,7,8,10)
c_melodic_minor <- c(0,2,3,5,7,9,11)
just_diatonic <- 12 * log2(c(1, 9/8, 5/4, 4/3, 3/2, 5/3, 15/8))
howfree(c_natural_minor)
howfree(c_melodic_minor)
howfree(just_diatonic)

}
