% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicMCT-package.R
\docType{package}
\name{musicMCT-package}
\alias{musicMCT}
\alias{musicMCT-package}
\title{musicMCT: Analysis of Musical Scale Structure}
\description{
Analysis of musical scales (as well as modes, grooves, etc.) in the vein of the author's article "Modal Color Theory." Offers support for conventional musical pitch class set theory (as developed by Allen Forte, David Lewin, etc.) as well as the continuous geometries of Callender, Quinn, & Tymoczko. Identifies structural properties of scales and calculates derived values (sign vector, color number, brightness ratio, etc.) as described in Sherrill 2025.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://satbq.github.io/musicMCT/}
}

}
\author{
\strong{Maintainer}: Paul Sherrill \email{paul.sherrill@utah.edu} (\href{https://orcid.org/0009-0002-3617-016X}{ORCID}) [copyright holder]

}
\keyword{internal}
