% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft.R
\name{set_to_distribution}
\alias{set_to_distribution}
\alias{s2d}
\title{Convert between pitch-class sets and distributions}
\usage{
set_to_distribution(set, edo = 12, rounder = 10)

s2d(...)
}
\arguments{
\item{set}{Numeric vector of pitch-classes in the set. May be a
multiset, in which case the result is different from the corresponding
set with repetitions removed. Entries must be integers.}

\item{edo}{Number of unit steps in an octave. Defaults to \code{12}.}

\item{rounder}{Numeric (expected integer), defaults to \code{10}:
number of decimal places to round to when testing for equality.}

\item{...}{Arguments to be passed from s2d() or d2s() to unabbreviated
functions.}
}
\value{
set_to_distribution() returns a numeric vector with length
\code{edo}, whose \code{i}th entry represents the weight assigned to pitch-class
\code{i} in the distribution. distribution_to_set() returns a (multi)set
represented by listing its elements. (Non-integer weights are rounded
\emph{up} to the next highest integer.)
}
\description{
For applications of the Discrete Fourier Transform to pitch-class
set theory, it's typically convenient to represent musical sets
in terms of \emph{distributions} rather than lists of their elements.
(See Chapter 1 of Amiot 2016, \url{doi:10.1007/978-3-319-45581-5}.)
These functions convert back and forth between those representations.
s2d() and d2s() are shorthands for set_to_distribution() and
distribution_to_set(), respectively.
}
\examples{
set_to_distribution(c(0, 4, 7))
s2d(c(0, 4, 7)) # Same result but quicker to type
s2d(c(0, 4, 4, 7)) # The doubled third is reflected by the value 2 in the result

}
