% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_scale.R
\name{simplify_scale}
\alias{simplify_scale}
\alias{best_simplification}
\title{Simplify a scale}
\usage{
simplify_scale(
  set,
  start_zero = TRUE,
  ineqmat = NULL,
  scales = NULL,
  signvector_list = NULL,
  adjlist = NULL,
  method = c("euclidean", "taxicab", "chebyshev", "hamming"),
  display_digits = 2,
  edo = 12,
  rounder = 10
)

best_simplification(set, ...)
}
\arguments{
\item{set}{Numeric vector of pitch-classes in the set}

\item{start_zero}{Boolean: should the result be transposed so that its pitch
initial is zero? Defaults to \code{TRUE}.}

\item{ineqmat}{Defaults to \code{NULL}, in which case the function assumes
you want to use the standard hyperplane arrangement of MCT. But you can
enter a different matrix that contains normal vectors for any hyperplane
arrangement in the same format as the standard ineqmats.}

\item{scales}{List of scales representing the faces of your hyperplane
arrangement. Defaults to \code{NULL} in which case the function looks for
\code{representative_scales} in the global environment.}

\item{signvector_list}{A list of signvectors to use as the reference by
which \code{colornum} assigns a value. Defaults to \code{NULL} and will attempt to
use \code{representative_signvectors}, which needs to be downloaded and assigned
separately from the package musicMCT.}

\item{adjlist}{Adjacency list structed in the same way as \code{color_adjacencies}.
Defaults to \code{NULL} in which case the function looks for \code{color_adjacencies}
in the global environment.}

\item{method}{What distance metric should be used? Defaults to \code{"euclidean"}
(unlike most functions with a method parameter in musicMCT)
but can be \code{"taxicab"}, \code{"chebyshev"}, or \code{"hamming"}.}

\item{display_digits}{Integer: how many digits to display when naming any
non-integral interval sizes. Defaults to 2.}

\item{edo}{Number of unit steps in an octave. Defaults to \code{12}.}

\item{rounder}{Numeric (expected integer), defaults to \code{10}:
number of decimal places to round to when testing for equality.}

\item{...}{Other arguments to be passed from \code{best_simplification()}
to \code{simplify_scale()}.}
}
\value{
A complicated matrix
}
\description{
simplify_scale description
}
\examples{
\dontrun{
  simplify_scale(c(0,5))
}
}
