% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_signvector.R
\name{step_signvector}
\alias{step_signvector}
\title{Specify a scale's step pattern with a sign vector}
\usage{
step_signvector(set, ineqmat = NULL, edo = 12, rounder = 10)
}
\arguments{
\item{set}{Numeric vector of pitch-classes in the set}

\item{ineqmat}{Defaults to \code{NULL}, in which case the function assumes
you want to use the standard hyperplane arrangement of MCT. But you can
enter a different matrix that contains normal vectors for any hyperplane
arrangement in the same format as the standard ineqmats.}

\item{edo}{Number of unit steps in an octave. Defaults to \code{12}.}

\item{rounder}{Numeric (expected integer), defaults to \code{10}:
number of decimal places to round to when testing for equality.}
}
\value{
A vectors of signs, \code{-1}, \code{0}, and \code{1}, corresponding to the step-related
hyperplanes in the defined \code{ineqmat}.
}
\description{
Rather than calculate the full sign vector from the "modal color" hyperplane
arrangement, sometimes it's advantageous to use a sign vector that reflects
only the pairwise comparisons on a scale's steps. This function does that.
}
\examples{
step_signvector(sc(7,35)) # Half the length of a full sign vector for heptachords:
signvector(sc(7,35))
}
