% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scala.R
\name{writeSCL}
\alias{writeSCL}
\title{Create a Scala tuning file from a given scale}
\usage{
writeSCL(x, filename, period = 2, ineqmat = NULL, edo = 12, rounder = 10)
}
\arguments{
\item{x}{Numeric vector: the scale to export}

\item{filename}{String (in quotation marks): what to name your Scala file. Will automatically
match the name of \code{x} if you enter nothing.}

\item{period}{The frequency ratio at which your scale repeats; defaults to \code{2} which
indicates an octave-repeating scale.}

\item{ineqmat}{Defaults to \code{NULL}, in which case the function assumes
you want to use the standard hyperplane arrangement of MCT. But you can
enter a different matrix that contains normal vectors for any hyperplane
arrangement in the same format as the standard ineqmats.}

\item{edo}{Number of unit steps in an octave. Defaults to \code{12}.}

\item{rounder}{Numeric (expected integer) representing tolerance:
number of decimal places to round to.}
}
\value{
Invisible \code{NULL}
}
\description{
You mean you don't want to play around in R forever?
This function lets you export any scale you've defined in R
as a .scl tuning file for use in Scala or by any synth that can read
.scl files. Will write to your working directory.

In addition to saving the necessary tuning data, the function
will attempt to add as comments extra information that can be derived
from MCT functions, like the color number, degrees of freedom, number of sign-vector zeroes, etc.
}
\examples{
neat_pentachord <- convert(c(0,1,4,9,11),15,12)
\dontrun{
  writeSCL(neat_pentachord, "neat_pentachord.scl")
}
}
