% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theory.R
\name{normal_form}
\alias{normal_form}
\title{Hook's OPTIC normal forms}
\usage{
normal_form(set, optic = "opc", edo = 12, rounder = 10)
}
\arguments{
\item{set}{Numeric vector of pitch-classes in the set}

\item{optic}{String: the OPTIC symmetries to apply. Defaults to "opc".}

\item{edo}{Number of unit steps in an octave. Defaults to \code{12}.}

\item{rounder}{Numeric (expected integer), defaults to \code{10}:
number of decimal places to round to when testing for equality.}
}
\value{
Numeric vector with the desired normal form of \code{set}
}
\description{
Following Hook (2023, 416-18, ISBN: 9780190246013), calculates a normal
form for the input \code{set} using any combination of OPTIC symmetries.
}
\examples{
5+5

}
