% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_VL.R
\name{tndists}
\alias{tndists}
\title{Distances between continuous transpositions of a set}
\usage{
tndists(set, method = c("taxicab", "euclidean"), subdivide = 100, edo = 12)
}
\description{
One way to think about the voice-leading potential of a set is to consider all the ways
that it can move to transpositions of itself. For instance, the major triad's closest
transpositions are \eqn{T_4} and \eqn{T_8} while its most distant transposition is \eqn{T_6},
and potentially also \eqn{T_{\pm 2}} depending on the distance metric you use. For
the major triad restricted to 12-tone equal temperament, this set of relationships is well
modeled by Richard Cohn's discussion of \href{https://doi.org/10.2307/843877}{Douthett & Steinbach's}
"Cube Dance" in \emph{Audacious Euphony} (102-106). The behavior of other sets is not always what you
might expect based on the model of tertian sonorities. For instance, the trichord (027) has
different minimal neighbors depending on the metric chosen: its nearest neighbors are \eqn{T_{\pm 4}}
under the Euclidean metric but \eqn{T_{\pm 5}} under the taxicab metric.
}
