% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theory.R
\name{tsym}
\alias{tsym}
\title{Test for transpositional symmetry}
\usage{
tsym(set, edo = 12, rounder = 10)
}
\arguments{
\item{set}{Numeric vector of pitch-classes in the set}

\item{edo}{Number of unit steps in an octave. Defaults to \code{12}.}

\item{rounder}{Numeric (expected integer), defaults to \code{10}:
number of decimal places to round to when testing for equality.}
}
\value{
\code{TRUE} if the set is transpositionally symmetrical, \code{FALSE} otherwise
}
\description{
Does the set map onto itself at some transposition other than \eqn{T_0}?
}
\examples{
tsym(sc(6, 34))
tsym(sc(6, 35))
tsym(edoo(5))

# Works for continuous values:
tsym(tc(j(dia), edoo(3)))

}
